mm: pagemap_read: limit scan to virtual region being asked
pagemap_read scans through the virtual address space of a task until it
prepares 'count' pagemaps or it reaches end of task.

This presents a problem when the page walk doesn't happen for a vma with
VM_PFNMAP set.  In this case the walk is silently skipped and no pagemap
is prepared, in turn making pagemap_read scan through to task end, even
crossing beyond 'count', landing into a different vma region.  This leads
to wrong presentation of mappings for that vma.

Fix this by limiting end_vaddr to the end of the virtual address region
being scanned.

Signed-off-by: Shiraz Hashim <shashim@codeaurora.org>
Cc: Oleg Nesterov <oleg@redhat.com>
Cc: Cyrill Gorcunov <gorcunov@openvz.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>