zsmalloc: do not remap dst page while prepare next src page
The object may belong to different pages.  zs_object_copy() handles this
case and maps a new source page (get_next_page() and kmap_atomic()) when
object crosses boundaries of the current source page.  But it also
performs unnecessary kunmap/kmap_atomic of the destination page (it
remains unchanged), which can be avoided.

Signed-off-by: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Minchan Kim <minchan@kernel.org>
Cc: Nitin Gupta <ngupta@vflare.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>